Refactor the code of accumulate add
