+++ chainer/function_node.py
@@ -8,8 +8,10 @@ import six
 
 import chainer
 from chainer.backends import cuda
+from chainer import _backprop_utils
 from chainer import configuration
 from chainer import function_hook
+from chainer.utils import experimental
 from chainer.utils import type_check
 from chainer import variable
 
@@ -310,6 +312,8 @@ Use apply() method instead.\
                 self._retained_output_data = tuple(retained_data)
 
             self.lazy_grad_sum = configuration.config.lazy_grad_sum
+            if self.lazy_grad_sum is True:
+                experimental('config.lazy_grad_sum')
 
         return ret
 
@@ -554,7 +558,7 @@ Use apply() method instead.\
         if self.lazy_grad_sum:
             gxs_output = ()
             for i, (gx, g_input) in enumerate(six.moves.zip(gxs, grad_inputs)):
-                sum_gx = chainer.functions.concat_variable(gx, g_input)
+                sum_gx = _backprop_utils.concat_variable(gx, g_input)
                 j = target_input_indexes[i]
                 if self.inputs[j].creator is None:
                     if sum_gx[0] is not None and len(sum_gx) > 1:
@@ -887,10 +891,10 @@ def _backprop(outputs, inputs, grad_required, retain_grad, grads):
                 cur_gx = grads.get(node, None)
                 if cur_gx is not None:
                     if x.creator is None:
-                        g = chainer.functions.concat_variable(g, cur_gx)
+                        g = _backprop_utils.concat_variable(g, cur_gx)
                         g = chainer.functions.accumulate_add(g)
                     else:
-                        g = chainer.functions.concat_variable(g, cur_gx)
+                        g = _backprop_utils.concat_variable(g, cur_gx)
             else:
                 selected_inputs.add(node)
