+++ chainer/variable.py
@@ -9,6 +9,7 @@ import numpy
 
 import chainer
 from chainer.backends import cuda
+from chainer import _backprop_utils
 from chainer import initializers
 from chainer.initializers import constant
 from chainer.utils import argument
@@ -1051,11 +1052,11 @@ Actual: {0}'''.format(type(data))
                     cur_gx = grads[x]
                     if func.lazy_grad_sum:
                         if x.creator is None:
-                            gx = chainer.functions.concat_variable(gx, cur_gx)
+                            gx = _backprop_utils.concat_variable(gx, cur_gx)
                             gx = chainer.functions.accumulate_add(gx)
                             grads[x] = gx
                         else:
-                            grads[x] = chainer.functions.concat_variable(
+                            grads[x] = _backprop_utils.concat_variable(
                                 gx, cur_gx)
                     else:
                         grads[x] = gx if cur_gx is None else gx + cur_gx
